% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment5.R
\docType{data}
\name{data}
\alias{data}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{reliability_amount}{}
\item{alignment}{}
\item{sex}{}
\item{age}{}
\item{id}{}
\item{project}{}
\item{npv_amount}{}
\item{forecast_mean}{}
\item{forecast_sd}{}
\item{sample}{}
}
}
\source{
The data was collected using Qualtrics.
}
\description{
Detailed description of the data
}
\seealso{
\link{alignment5}
}
