% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getforecast.R
\name{getforecast}
\alias{getforecast}
\title{Get mean and SD of forecasts}
\usage{
getforecast(data, forecastName, fun, weightings)
}
\arguments{
\item{data}{a dataframe.}

\item{forecastName}{name of relevant DV.}

\item{fun}{}

\item{weightings}{distribution forecast weightings.}
}
\value{
A dataframe.
}
\description{
Get mean and SD of forecasts
}
\examples{
\dontrun{
getforecast(data, forecastName, weightings)
}
}
