% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmean.R
\name{getmean}
\alias{getmean}
\title{Get mean for forecast data}
\usage{
getmean(df, weightings)
}
\arguments{
\item{df}{a dataframe.}

\item{weightings}{distribution forecast weightings.}
}
\value{
forecast means.
}
\description{
Get mean for forecast data
}
\examples{
\dontrun{
getmean(df, weightings)
}
}
