% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsd.R
\name{getsd}
\alias{getsd}
\title{Get SD for forecast data}
\usage{
getsd(df, weightings)
}
\arguments{
\item{df}{a dataframe.}

\item{weightings}{distribution forecast weightings.}
}
\value{
forecast SDs.
}
\description{
Get SD for forecast data
}
\examples{
\dontrun{
getsd(df, weightings)
}
}
