% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment5.R
\docType{package}
\name{alignment5}
\alias{alignment5}
\alias{alignment5-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
alignment5
A data package for alignment5.

Use \code{data(package='alignment5')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
